Set environment variable containing path to index.html
When we load index.html via a non-file: uri, electron's renderer/init.js can't find the file, so it can't set up the module search paths correctly, and require'ing dependencies from a renderer process fails.

This change sets an environment variable with the path to index.html that the renderer process can read and patch up the module search paths to be correct.